refactor: WeekCell 폴더 분리 및 컴포넌트명 타이포 수정
