+++ frontend/src/components/Calendar/Core/index.tsx
@@ -1,9 +1,9 @@
-import { SelectedWeak } from './SelectedWeak';
+import { SelectedWeek } from './SelectedWeek';
 import { TimeControl } from './TimeControl';
 
 export const Core = () => (
   <>
     <TimeControl />
-    <SelectedWeak />
+    <SelectedWeek />
   </>
 );
